requirements.txt
reportlab>=4.0.0
pyyaml>=6.0
anthropic>=0.25.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
